Further widgetization cleanup and background-issue debugging.
This cleans up the worst of the remaining issues left over from
widgetization, including the big perf issue and two bugs that
were both preventing diffuse-background from working correctly.
One was an issue with the egg game's main widget not getting
real transparency because of a bug in blankrect's handling of
callbacks.  The was the diffuse widget's color handling, which
had not been brought up to date with the way color is now
(supposed to be) handled in widgets.pl